backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
distlib==0.3.4
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
filelock==3.4.2
importlib-metadata==4.11.1
Markdown==3.3.6
Pillow==9.0.1
platformdirs==2.5.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.13.1
zipp==3.7.0
